% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{cc2plot}
\alias{cc2plot}
\title{Plot a Control Chart.}
\usage{
cc2plot (data, data.title = "")
}
\arguments{
\item{data}{A cchart object. See ?ccharter::cchart function for reference.}

\item{data.title}{Title for the Control Chart plot.}
}
\description{
Plots a Control Chart. Receives as input a cchart object.
}
\examples{
# Create a random time series
set.seed(154)
time.series <- data.frame(t.dates = seq.Date(as.Date("2014-02-01"), as.Date("2016-08-01"), "month"),
               t.values = c(
               seq(0.1, 0.8, by = 0.1) * runif(8) + 3,
               seq(0.1, 0.7, by = 0.1) * runif(7) + 4,
               seq(0.1, 0.7, by = 0.1) * runif(7) + 5,
               seq(0.1, 0.4, by = 0.1) * runif(4) + 4,
               seq(0.1, 0.5, by = 0.1) * runif(5) + 4)
               )

# Execute function
control.chart.data <- cchart(time.series, "t.dates", "t.values")

# Create chart
cc2plot(control.chart.data)


}

