% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchart.R
\name{ccpoints}
\alias{ccpoints}
\title{Calculates Control Systems points}
\usage{
ccpoints(data, dates, values, points.vs.avg = 7,
         date.type = TRUE, already.ordered = FALSE)
}
\arguments{
\item{data}{The input Data Frame.}

\item{dates}{Column name for date or ordering variable.}

\item{values}{Column name for data points values. Class should be numeric,
if not, will attempt to coerce.}

\item{points.vs.avg}{For Mean Rule, establishes how many continuous points
should be above or below mean to define a system. Default to 6.}

\item{date.type}{States if observations are based on dates. Defaults to TRUE. If TRUE, will attempt to coerce
to a date format using extensive use of lubridate package dates formats.}

\item{already.ordered}{Tells the function if the data is already ordered. Defaults to FALSE.
The function will attempt to order it.}
}
\value{
The function will return a list object with class ccpoints.
A list including the data frame with the system points, and column names of the series.\cr
The data frame will include the submited series and the following new columns: \cr\cr
data.mean: the mean value for each system \cr
data.ll: the lower standard deviation of each system \cr
data.ul: the upper standard deviation of each system \cr
}
\description{
Control Charts made easy.
Calculates Control Systems for a time series.
The current distribution uses the following Control Rules to define systems:
1. Mean Rule: N continuous points above or below mean. See points.vs.avg argument.
2. Out of Control Points Rule: Points above or below 3 Standard Deviation are excluded from calculations.
}
\examples{
# Create a random time series
set.seed(154)
time.series <- data.frame(t.dates = seq.Date(as.Date("2014-02-01"), as.Date("2016-08-01"), "month"),
               t.values = c(
               seq(0.1, 0.8, by = 0.1) * runif(8) + 3,
               seq(0.1, 0.7, by = 0.1) * runif(7) + 4,
               seq(0.1, 0.7, by = 0.1) * runif(7) + 5,
               seq(0.1, 0.4, by = 0.1) * runif(4) + 4,
               seq(0.1, 0.5, by = 0.1) * runif(5) + 4)
               )

# Execute function
control.chart.data <- ccpoints(time.series, "t.dates", "t.values")
print(control.chart.data)

# To extract only the data frame
control.chart.data[["data"]]

}

